+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Add `Transformed` element](https://github.com/square/Blueprint/pull/195) to apply a `CATransform3D` to a wrapped element.
+
 ### Removed
 
 ### Changed